Auth controller added, endpoint fixed, password response fixed - Changes on login
