+++ app.js
@@ -7,7 +7,7 @@ const cors = require("cors");
 require("dotenv").config();
 
 const indexRouter = require("./routes/index");
-const newsRouter = require("./routes/auth");
+const authRouter = require("./routes/auth");
 const usersRouter = require("./routes/users");
 const rolesRouter = require("./routes/roles");
 const testimonialsRouter = require("./routes/testimonials");
@@ -15,8 +15,7 @@ const membersRouter = require("./routes/members");
 const categoriesRouter = require("./routes/categories");
 const organizationRouter = require("./routes/organization");
 const newsRouter = require("./routes/news");
-const loginRouter = require('./routes/login');
-const { swaggerJsDoc, swaggerUi, swaggerSpec} = require("./services/swagger");
+const { swaggerJsDoc, swaggerUi, swaggerSpec } = require("./services/swagger");
 const app = express();
 app.use(cors());
 
@@ -29,8 +28,8 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
-app.use('/auth/login', loginRouter);
-app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(swaggerJsDoc(swaggerSpec)));
+app.use("/auth", authRouter);
+app.use("/api/docs", swaggerUi.serve, swaggerUi.setup(swaggerJsDoc(swaggerSpec)));
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/roles", rolesRouter);
