+++ controllers/users.js
@@ -1,4 +1,6 @@
 const { User } = require("../models");
+// Nano: Import bcrypt library to encrypt passwords
+const bcrypt = require("bcrypt");
 
 const getUsers = async (req, res) => {
   try {
@@ -25,9 +27,20 @@ const getUser = async (req, res) => {
 const createUser = async (req, res) => {
   try {
     const { firstName, lastName, email, password, photo, roleId } = req.body;
-    const user = await User.create({ firstName, lastName, email, password, photo, roleId });
+    // Nano: Create salt and make hash to encrypt passwords
+    const salt = await bcrypt.genSalt();
+    const encryptedPassword = await bcrypt.hash(password, salt);
+    console.log({ firstName, lastName, email, encryptedPassword, photo, roleId });
+    const user = await User.create({
+      firstName,
+      lastName,
+      email,
+      password: encryptedPassword,
+      photo,
+      roleId,
+    });
     user.save();
-    return res.status(201).json(user);
+    return res.status(201).json({ ...user.dataValues, password: undefined });
   } catch (error) {
     return res.status(500).json(error);
   }
